backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.29.0
pydantic==2.4.2
python-dotenv==1.0.0